Add Home, Login Screens
Add Home, Login Screens. Add buttons and functionality to move between pages.
